Move rpm-md parsing into a shared module
We also need to parse repomd.xml + primary.xml for other distributions.
Therefore we move it to schedule::repomd to allow reuse.

The fedora.rs keeps the directory layout of the mirror and how
the packages are grouped into reports.
No functional change, the sync output is unchanged.
